Fix merging values for recurring category names
Some category names are duplicated in different files (i.e. 'games'),
therefore the value for the key 'games' was overwritten each time a new file was read.

This commit fixes that by doing a check if the key with a given name is already present in
the map and then merges the values.