Update native-image config files
Apart from adding reflection config for new providers had to add an
entry for `kotlin.jvm.functions.Function0` since native image was
throwing an exception with `--verbose` option that this class is not
found.

native-image agentlib also generated a jni-config.json file this
time, which I tried building with as well.
Could be that it was the one causing the above exception.