Numerify/letterify values after they have been resolved
Fix #56

This makes it easier to work with, as we don't need to create
comlex regexes, check different conditions, and so on.

This also greatly simplifies working with `letterify` function,
since previously we were checking for '??' in the regex
because some strings contained '?'
but were not supposed to be letterified (i.e. quotes)