Add missing 'fr.yml' dict file
This file is part of locales in ruby-faker and was accidentally
removed from kotlin-faker in one of previous commits.

It's not used because 'fr/phone_number.yml' also contains the same
key:value pair, but still doesn't hurt to commit it to avoid the
"new file noise" when merging dict from upstream.