Add 'release' make target that accepts version string
A bit annoying that I depend on a local version of
semantic-release-versioning plugin, hence the new make target allows
to set the version via gradle property instead of depending on the
plugin.

This is hopefully temporary (as in - "the most permanent things begin as temporary ones")
until I have time to get my own semver gradle plugin up and running.