Fix FakerService#getProvider for standalone provider implementations
BookFaker could try to resolve "#{Name.name}" in one of its functions,
which would try to lookup the `Name` provider that isn't available
in the BooksFaker.

In such cases we try to find the `Name` provider in the core `Faker`
implementation, and failing that thrown an exception to make the
error clearer.

We do not try to support CustomFaker -> AnotherCustomFaker calls,
e.g. BooksFaker -> MusicFaker.
Only CustomFaker -> Faker is supported, i.e. BooksFaker -> Faker.