Change visibility of AbstractFakeDataProvider#fakerService to protected
Relates to changes introduced in #219, particularly fakerService
became publicly visible, which is not ideal.

To hide it we had to introduce another property in
UniqueProviderDelegate, since that was the only place where
compilation was failing. So instead of using fakerService from
thisRef, we make it constructor property and use it from there.

This of course also touched almost all data providers.

Not sure if this needs more testing though.