Make 'docs' a standalone module
Docs are finally starting to fail hard...

Version 0.21.1 of Orchid seems to no longer be available:

 Could not resolve all dependencies for configuration ':docs:detachedConfiguration3'.
   > Could not find io.github.javaeden.orchid:OrchidCore:0.21.1.
     Searched in the following locations:
       - https://repo.maven.apache.org/maven2/io/github/javaeden/orchid/OrchidCore/0.21.1/OrchidCore-0.21.1.pom
       - https://plugins.gradle.org/m2/io/github/javaeden/orchid/OrchidCore/0.21.1/OrchidCore-0.21.1.pom
     Required by:
         project :docs

I've discovered that there's a version 1.0.0-SNAPSHOT published to maven
central, and tried using that, but it also fails:

[INFO] GeneratorServiceImpl:     Generating [indices]
Error: Exception in thread "main" java.lang.NoSuchMethodError: 'double kotlin.time.DurationKt.toDuration(long, java.util.concurrent.TimeUnit)'
	at com.eden.orchid.api.generators.GeneratorMetrics.getIndexingTime-UwyO8pc$orchid_core(GeneratorMetrics.kt:28)

This seems to be due to breaking changes in kotlin, since
orchid:1.0.0-SNAPSHOT uses 1.4.31 and we're using 1.9.x in kotlin-faker currently.

So the only option I could see before #166 is done (which I still can't
find time for...), taking docs out of the kotlin-faker build
configuration and converting it to a standalone gradle project, with own
build configs, a lower gradle and kotlin version (due to
kotlin-to-gradle compatibilities), and so on, seems like the only viable
option.

The good part of this whole mess is that this workaround
actually seems to be working... phew.